requirements.txt
tensorflow-gpu~=1.14.0
pandas~=1.0.3
numpy~=1.16.4
matplotlib~=3.1.1
psycopg2~=2.8.4
scipy~=1.3.0
scikit-learn~=0.21.3